backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
sqlmodel>=0.0.14
psycopg[binary]>=3.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0
alembic>=1.13.0
pyjwt>=2.8.0
argon2-cffi>=23.1.0
email-validator>=2.0.0
